requirements.txt
asgiref==3.7.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
crispy-bootstrap4==2022.1
cryptography==41.0.1
Django==4.2.2
django-crispy-forms==2.0
django-rest-knox==4.2.0
djangorestframework==3.14.0
idna==3.4
mysqlclient==2.1.1
pycparser==2.21
pytz==2023.3
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.6.3
tzdata==2023.3
urllib3==2.0.2
